requirements.txt
cryptography==3.4.8
easygui==0.98.3
python-dotenv==1.0.0
